Create README.md
Fast and accurate intelligent classification of rocks is the most important prerequisite to ensure the unmanned construction of geological projects. In this paper, a dataset containing 660 images is established, classified based on the structure and the fabric characteristics of rocks, and investigate the recognition accuracy of different structural and tectonic features. Moreover, the accuracy, training time, and robustness of different neural network architectures on a multi-scale image set are compared by transfer learning. The results show that InceptionResnetV2 has the best robustness and high accuracy, which can reach 74.24%, 74.24%, and 72.73% on the three image sets respectively; and it is found that the Inception module significantly improves the recognition accuracy and the Fire module reduces the learning time by a factor of 5–10. Finally, MobileNetV2 was tested and it obtained the highest accuracy of 81.82%. In addition, MobileNetV2 has a strong ability to recognize typical rock structures such as bedding, columnar joints, and leaf veins, and can achieve a recognition accuracy of 95.83%.